requirements.txt
pycryptodome
cryptography
tqdm
pythonping